Mise à jour des contrôleurs, modèles et autres fichiers

+++ app/Http/Controllers/CategoriesController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 use App\Models\Categories;
+use Illuminate\Support\Facades\Auth;
 
 use Illuminate\Http\Request;
 
 
     class CategoriesController extends Controller
+
     {
+
+
+
+
         // Afficher toutes les catégories
         public function index()
         {
@@ -24,14 +30,24 @@ use Illuminate\Http\Request;
 
         // Enregistrer une nouvelle catégorie
         public function store(Request $request)
-        {
+        {     $user = Auth::user();
+            $roles = ['admin', 'superadmin', 'dispatcheur', 'operateur', 'responsable_marketing'];
+
+            if (!$user || !$user->hasAnyRole($roles)) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
             $categorie = Categories::create($request->all());
             return response()->json($categorie, 201);
         }
 
         // Mettre à jour une catégorie
         public function update(Request $request, $id)
-        {
+        {     $user = Auth::user();
+            $roles = ['admin', 'superadmin', 'dispatcheur', 'operateur', 'responsable_marketing'];
+
+            if (!$user || !$user->hasAnyRole($roles)) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
             $categorie = Categories::findOrFail($id);
             $categorie->update($request->all());
             return response()->json($categorie, 200);
@@ -39,7 +55,12 @@ use Illuminate\Http\Request;
 
         // Supprimer une catégorie
         public function destroy($id)
-        {
+        {     $user = Auth::user();
+            $roles = ['admin', 'superadmin', 'dispatcheur', 'operateur', 'responsable_marketing'];
+
+            if (!$user || !$user->hasAnyRole($roles)) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
             Categories::findOrFail($id)->delete();
             return response()->json(null, 204);
         }
